+++ src/utils/apiEndpoints.ts
@@ -10,6 +10,7 @@ export const NetworkToIndexerAPI: Record<string, string> = {
   testnet: "https://api.testnet.aptoslabs.com/v1/graphql",
   devnet: "https://api.devnet.aptoslabs.com/v1/graphql",
   shelbynet: "https://api.shelbynet.shelby.xyz/v1/graphql",
+  netna: "https://api.netna.staging.aptoslabs.com/v1/graphql",
   local: "http://127.0.0.1:8090/v1/graphql",
 };
 
@@ -22,6 +23,7 @@ export const NetworkToNodeAPI: Record<string, string> = {
   testnet: "https://api.testnet.aptoslabs.com/v1",
   devnet: "https://api.devnet.aptoslabs.com/v1",
   shelbynet: "https://api.shelbynet.shelby.xyz/v1",
+  netna: "https://api.netna.staging.aptoslabs.com/v1",
   local: "http://127.0.0.1:8080/v1",
 };
 
@@ -32,6 +34,7 @@ export const NetworkToNodeAPI: Record<string, string> = {
 export const NetworkToFaucetAPI: Record<string, string> = {
   devnet: "https://faucet.devnet.aptoslabs.com",
   shelbynet: "https://faucet.shelbynet.shelby.xyz",
+  netna: "https://faucet-dev-netna-us-central1-410192433417.us-central1.run.app",
   local: "http://127.0.0.1:8081",
 };
 
@@ -44,6 +47,7 @@ export const NetworkToPepperAPI: Record<string, string> = {
   testnet: "https://api.testnet.aptoslabs.com/keyless/pepper/v0",
   devnet: "https://api.devnet.aptoslabs.com/keyless/pepper/v0",
   shelbynet: "https://api.shelbynet.aptoslabs.com/keyless/pepper/v0",
+  netna: "https://api.devnet.aptoslabs.com/keyless/pepper/v0",
   // Use the devnet service for local environment
   local: "https://api.devnet.aptoslabs.com/keyless/pepper/v0",
 };
@@ -57,6 +61,7 @@ export const NetworkToProverAPI: Record<string, string> = {
   testnet: "https://api.testnet.aptoslabs.com/keyless/prover/v0",
   devnet: "https://api.devnet.aptoslabs.com/keyless/prover/v0",
   shelbynet: "https://api.shelbynet.aptoslabs.com/keyless/prover/v0",
+  netna: "https://api.devnet.aptoslabs.com/keyless/prover/v0",
   // Use the devnet service for local environment
   local: "https://api.devnet.aptoslabs.com/keyless/prover/v0",
 };
@@ -71,6 +76,7 @@ export enum Network {
   TESTNET = "testnet",
   DEVNET = "devnet",
   SHELBYNET = "shelbynet",
+  NETNA = "netna",
   LOCAL = "local",
   CUSTOM = "custom",
 }
@@ -94,6 +100,7 @@ export const NetworkToNetworkName: Record<string, Network> = {
   testnet: Network.TESTNET,
   devnet: Network.DEVNET,
   shelbynet: Network.SHELBYNET,
+  netna: Network.NETNA,
   local: Network.LOCAL,
   custom: Network.CUSTOM,
 };